fix: code review fixes for PR #76 and iOS 26 test infrastructure
- AdversaryRunnerCard: add @FocusState for immediate keyboard focus on
  rename, add AX identifiers to Done/Cancel buttons, clean up nameError init
- AdversaryRunnerSection: include active conditions in collapsed row
  accessibilityLabel (PR #75 broke this by switching from accessibilityValue
  to an explicit label that omitted condition state)
- EncounterUITestCase: new shared base class for RunnerUITests and
  RenameAdversaryUITests, eliminating ~180 lines of duplicated helpers
- Navigation helpers: fix iOS 26 toolbar overflow — builder toolbar items
  collapse into OverflowBarButtonItem; use builder.add-adversary-button
  (in-form) for compendium and tap overflow before "Run Encounter"
- replaceText: remove redundant initial tap on iOS; triple-tap alone
  handles both pre-focused (@FocusState) and unfocused states correctly

All 9 iOS UI tests and all macOS unit tests pass.
